+++ Tests/Integration/Microsoft365DSC/M365DSCIntegration.AAD.Remove.Tests.ps1
@@ -112,6 +112,23 @@
                     Ensure                = "Absent";
                     Id                    = "HardwareOath";
                 }
+                AADAuthenticationMethodPolicyQRCodeImage 'AADAuthenticationMethodPolicyQRCodeImage-QRCodePin'
+                {
+                    ApplicationId                = $ApplicationId;
+                    CertificateThumbprint        = $CertificateThumbprint;
+                    Ensure                       = "Absent";
+                    Id                           = "QRCodePin";
+                    IncludeTargets               = @(
+                        MSFT_AADAuthenticationMethodPolicyQRCodeImageIncludeTarget{
+                            Id = "all_users"
+                            TargetType = "group"
+                        }
+                    );
+                    PinLength                    = 9; # Drift
+                    StandardQRCodeLifetimeInDays = 365;
+                    State                        = "disabled";
+                    TenantId                     = $TenantId;
+                }
                 AADAuthenticationMethodPolicySms 'AADAuthenticationMethodPolicySms-Sms'
                 {
                     ApplicationId         = $ApplicationId