+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.AzureBillingAccountScheduledAction.Tests.ps1
@@ -221,7 +221,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             }
 
             It 'Should return true from the Test method' {
-                Test-TargetResource @testParams | Should -Be $true
+                Test-TargetResource @testParams -verbose | Should -Be $true
             }
         }