+++ CHANGELOG.md
@@ -1,14 +1,37 @@
 # Change log for Microsoft365DSC
 
+# UNRELEASED
+
+* AADAuthenticationMethodPolicyQRCodeImage
+  * Initial release.
+* AADRoleEligibilityScheduleRequest
+  * Updated startdatetime and enddatetime properties to be of type DATETIME.
+* IntuneDeviceEnrollmentPlatformRestriction
+  * Added note that update is no longer possible.
+    FIXES [#5127](https://github.com/microsoft/Microsoft365DSC/issues/5127)
+* O365SearchAndIntelligenceConfigurations
+  * Added note that only Credentials are supported for the resource.
+* PPEnvironmentAppsEnvironment
+  * Added support for legacy Basic and Standard EnvironmentSKU.
+* TeamsOrgWideAppSettings
+  * Added note that only Credentials are supported for the resource.
+    FIXES [#3394](https://github.com/microsoft/Microsoft365DSC/issues/3394)
+* M365DSCDRGUtil
+  * Fixes an issue with fancy double quotes being replaced that break the string.
+    FIXES [#5775](https://github.com/microsoft/Microsoft365DSC/issues/5775)
+    FIXES [#5623](https://github.com/microsoft/Microsoft365DSC/issues/5623)
+
 # 1.25.226.1
 
 * AADConditionalAccessPolicy
-  * Fixed an issue where `TermsOfUse` was not passed as an array, causing failures in GCC-High environments.
+  * Fixed an issue where `TermsOfUse` was not passed as an array, causing
+    failures in GCC-High environments.
     FIXES [#5742](https://github.com/microsoft/Microsoft365DSC/issues/5742)
   * Added verbose to the Get-TargetResource function to print out the retrieved
     policies from calling the cmdlet.
 * AADPasswordRuleSettings
-  * Updated schema to only accept values 'Enforced' and 'Audit' for parameter BannedPasswordCheckOnPremisesMode
+  * Updated schema to only accept values 'Enforced' and 'Audit' for parameter
+    BannedPasswordCheckOnPremisesMode
 * IntuneDeviceCompliancePolicyWindows10
   * Fixes the handling of the `DeviceCompliancePolicyScript` property.
     FIXES [#5510](https://github.com/microsoft/Microsoft365DSC/issues/5510)